feat: Make package PEP conformant and PyPI ready
This commit addresses several items to improve the package quality and prepare it for PyPI submission.

- **Cleaned up repository structure:** Moved all project files to the root directory for a standard layout.
- **Updated author information:** Set the author to Gowtham Rao <rao@ohdsi.org> in `pyproject.toml`.
- **Added NOTICE file:** Created a `NOTICE` file with copyright information.
- **PEP 8 conformance:** Fixed all automatically and manually fixable PEP 8 issues reported by `ruff`. Suppressed complexity warnings for two functions that would require significant refactoring.
- **PyPI readiness:** Updated `pyproject.toml` with a description, MIT license, and classifiers. Replaced the existing Apache 2.0 license with an MIT license to match `pyproject.toml`.
