+++ docs/adrs/002-use-pydantic-for-validation-and-settings.md
