--- py-load-epar/src/py_load_epar/spor_api/__init__.py
@@ -1,1 +1,0 @@
-# This file makes the spor_api directory a Python package.
