+++ src/py_load_epar/models.py
@@ -2,7 +2,7 @@ import datetime
 from typing import Optional
 from uuid import UUID
 
-from pydantic import BaseModel, Field, ConfigDict
+from pydantic import BaseModel, ConfigDict, Field
 
 
 class PipelineExecution(BaseModel):
