+++ tests/etl/test_downloader.py
@@ -1,73 +1,59 @@
-import io
-import logging
+import hashlib
+from unittest.mock import MagicMock
 
 import pytest
-import requests
-import requests_mock
-from py_load_epar.etl.downloader import download_file, parse_excel_data
 
-# Configure logging for tests
-logging.basicConfig(level=logging.INFO)
+from py_load_epar.etl.downloader import (
+    _download_file_to_stream,
+    download_document_and_hash,
+)
+from py_load_epar.storage.interfaces import IStorage
 
 
-def test_download_file_success(requests_mock):
+def test_download_document_and_hash_with_mock_storage(requests_mock):
     """
-    Tests that download_file successfully downloads content on a 200 OK response.
+    Test that a document is downloaded, hashed, and saved via a mock storage adapter.
     """
-    url = "http://test.com/data.xlsx"
-    mock_content = b"dummy excel content"
-    requests_mock.get(url, content=mock_content, status_code=200)
+    url = "https://fake-ema-url.com/document.pdf"
+    mock_content = b"some_pdf_content_for_testing"
+    requests_mock.get(url, content=mock_content)
 
-    result_buffer = download_file(url)
+    # Create a mock storage object that adheres to the IStorage interface
+    mock_storage = MagicMock(spec=IStorage)
+    expected_uri = "mock://storage/documents/document.pdf"
+    mock_storage.save.return_value = expected_uri
 
-    assert isinstance(result_buffer, io.BytesIO)
-    assert result_buffer.read() == mock_content
+    # Calculate the expected hash
+    expected_hash = hashlib.sha256(mock_content).hexdigest()
+
+    # Call the function with the mock storage
+    storage_uri, file_hash = download_document_and_hash(url, mock_storage)
+
+    # --- Assertions ---
+    # 1. Check the return values are correct
+    assert storage_uri == expected_uri
+    assert file_hash == expected_hash
+
+    # 2. Check that the 'save' method on the mock storage was called correctly
+    mock_storage.save.assert_called_once()
+
+    # 3. Inspect the arguments passed to the 'save' method
+    call_args = mock_storage.save.call_args
+    saved_stream = call_args.kwargs['data_stream']
+    saved_object_name = call_args.kwargs['object_name']
+
+    assert saved_stream.read() == mock_content
+    assert saved_object_name == "documents/document.pdf"
 
 
-def test_download_file_http_error(requests_mock):
-    """
-    Tests that download_file raises a RequestException on an HTTP error (e.g., 404).
-    """
-    url = "http://test.com/notfound"
-    requests_mock.get(url, status_code=404, reason="Not Found")
+def test_download_raises_exception_on_http_error(requests_mock):
+    """Test that the downloader raises an exception for a 404 error."""
+    url = "https://fake-ema-url.com/not_found"
+    requests_mock.get(url, status_code=404)
 
-    with pytest.raises(requests.exceptions.HTTPError) as excinfo:
-        download_file(url)
+    with pytest.raises(Exception):
+        # Use a dummy file stream for the test
+        from io import BytesIO
 
-    assert "404 Client Error: Not Found" in str(excinfo.value)
-
-
-def test_parse_excel_data():
-    """
-    Tests that parse_excel_data correctly parses an in-memory Excel file.
-    """
-    # 1. Create an in-memory Excel file
-    output = io.BytesIO()
-    workbook = openpyxl.Workbook()
-    sheet = workbook.active
-    sheet.title = "Test Data"
-
-    # 2. Add data
-    header = ["ID", "Medicine Name", "Value"]
-    rows = [
-        (1, "TestMed A", 100),
-        (2, "TestMed B", 200),
-    ]
-    sheet.append(header)
-    for row in rows:
-        sheet.append(row)
-
-    # 3. Save to buffer
-    workbook.save(output)
-    output.seek(0)  # Rewind the buffer to the beginning
-
-    # 4. Parse the data
-    parsed_records = parse_excel_data(output)
-
-    # 5. Check the results
-    assert hasattr(parsed_records, '__iter__')  # Check if it's an iterator
-    result_list = list(parsed_records)
-
-    assert len(result_list) == 2
-    assert result_list[0] == {"ID": 1, "Medicine Name": "TestMed A", "Value": 100}
-    assert result_list[1] == {"ID": 2, "Medicine Name": "TestMed B", "Value": 200}
+        dummy_stream = BytesIO()
+        _download_file_to_stream(url, dummy_stream)
